+++ ios/Classes/AudioGraphBuilderPlugin.swift
@@ -65,8 +65,8 @@ public class AudioGraphBuilderPlugin: NSObject, FlutterPlugin {
     
     private func createEngineNode(_ node: AudioNode, engine: AVAudioEngine, inputConnections: [AudioNodeConnection], outputConnections: [AudioNodeConnection]) throws -> AudioEngineNode? {
         switch node.name {
-        case AudioEngineFileNode.nodeName:
-            return try AudioEngineFileNode(node, inputConnections: inputConnections, outputConnections: outputConnections)
+        case AudioEngineFilePlayerNode.nodeName:
+            return try AudioEngineFilePlayerNode(node, inputConnections: inputConnections, outputConnections: outputConnections)
         case AudioEngineMixerNode.nodeName:
             return try AudioEngineMixerNode(node, inputConnections: inputConnections, outputConnections: outputConnections)
         case AudioEngineDeviceOutputNode.nodeName:
